# docs/openapi-Khipu.json
{
  "openapi": "3.1.0",
  "info": {
    "license": {
      "name": "API Pagos Instantáneos",
      "url": "https://khipu.com"
    },
    "title": "API Pagos Instantáneos",
    "description": "La API Instant Payment de Khipu (versión 3.0) permite a comerciantes iniciar, consultar, confirmar,  reembolsar o eliminar pagos inmediatos mediante transferencias bancarias en tiempo real.  Entre sus principales funcionalidades se encuentran: obtener la lista de bancos disponibles  con detalles como límites, tarifas y logos; crear un pago especificando un banco o dejar que el  usuario elija; confirmar o solicitar el reembolso de un pago mediante su ID; y predecir el estado  de un pago usando machine learning. Esta API maneja autenticación mediante clave secreta en el  encabezado x-api-key, no requiere firmas, y utiliza certificados SSL de validación extendida para  asegurar la comunicación",
    "version": "v3.0",
    "x-logo": {
      "url": "https://s3.amazonaws.com/static.khipu.com/2021/logo-transparent.png",
      "altText": "Logo de Khipu",
      "href": "https://khipu.com"
    }
  },
  "servers": [
    {
      "url": "https://payment-api.khipu.com",
      "description": "Producción"
    }
  ],
  "security": [
    {
      "Api-Key": []
    }
  ],
  "paths": {
...
    "/v3/payments/{id}": {
      "get": {
        "operationId": "getPaymentById",
        "summary": "Get payment by Id",
        "description": "Información completa del pago. Datos con los que fue creado y el estado actual del pago.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Identificador del pago"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa. Devuelve un objeto JSON con la información de pago.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/payment-get-payment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Solicitud mal formada"
          },
          "401": {
            "description": "Unauthorized - Credenciales no válidas o faltantes"
          },
          "403": {
            "description": "Forbidden - Cliente no tiene privilegios"
          },
          "404": {
            "description": "Not found - Recurso no disponible"
          },
          "500": {
            "description": "Internal server error - Servidor no pudo completar la solicitud"
          }
        }
      },
      "delete": {
        "operationId": "deletePaymentById",
        "summary": "Delete payment by Id",
        "description": "Borrar un pago. Solo se pueden borrar pagos que estén pendientes de pagar. Esta operación no puede deshacerse.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "Identificador del pago"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Solicitud mal formada"
          },
          "401": {
            "description": "Unauthorized - Credenciales no válidas o faltantes"
          },
          "403": {
            "description": "Forbidden - Cliente no tiene privilegios"
          },
          "404": {
            "description": "Not found - Recurso no disponible"
          },
          "500": {
            "description": "Internal server error - Servidor no pudo completar la solicitud"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Api-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "success": {
        "type": "object",
        "description": "Objeto para respuestas exitosas de la API.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensaje a desplegar al usuario.",
            "example": "Message."
          }
        }
      },
      "payment-get-payment": {
        "type": "object",
        "description": "Información completa del pago. Datos con los que fue creado y el estado actual del pago.",
        "required": [
          "payment_id",
          "payment_url",
          "simplified_transfer_url",
          "transfer_url",
          "app_url",
          "ready_for_terminal",
          "notification_token",
          "receiver_id",
          "conciliation_date",
          "subject",
          "amount",
          "currency",
          "status",
          "status_detail",
          "body",
          "picture_url",
          "receipt_url",
          "return_url",
          "cancel_url",
          "notify_url",
          "notify_api_version",
          "expires_date",
          "attachment_urls",
          "bank",
          "bank_id",
          "payer_name",
          "payer_email",
          "personal_identifier",
          "bank_account_number",
          "out_of_date_conciliation",
          "transaction_id",
          "custom",
          "responsible_user_email",
          "send_reminders",
          "send_email",
          "payment_method",
          "funds_source"
        ],
        "properties": {
          "payment_id": {
            "type": "string",
            "description": "Identificador único del pago, es una cadena alfanumérica de 12 caracteres. Como este identificador es único, se puede usar, por ejemplo, para evitar procesar una notificación repetida. (Khipu espera un código 200 al notificar un pago, si esto no ocurre se reintenta hasta por dos días).",
            "example": "gqzdy6chjne9"
          },
          "payment_url": {
            "type": "string",
            "description": "URL principal del pago, si el usuario no ha elegido previamente un método de pago se le muestran las opciones.",
            "example": "https://khipu.com/payment/info/gqzdy6chjne9"
          },
          "simplified_transfer_url": {
            "type": "string",
            "description": "URL de pago simplificado.",
            "example": "https://app.khipu.com/payment/simplified/gqzdy6chjne9"
          },
          "transfer_url": {
            "type": "string",
            "description": "URL de pago normal.",
            "example": "https://khipu.com/payment/manual/gqzdy6chjne9"
          },
          "app_url": {
            "type": "string",
            "description": "URL para invocar el pago desde un dispositivo móvil usando la APP de Khipu.",
            "example": "khipu:///pos/gqzdy6chjne9"
          },
          "ready_for_terminal": {
            "type": "boolean",
            "description": "Es `true` si el pago ya cuenta con todos los datos necesarios para abrir directamente la aplicación de pagos Khipu.",
            "example": false
          },
          "notification_token": {
            "type": "string",
            "description": "Cadena de caracteres alfanuméricos que identifican unicamente al pago, es el identificador que el servidor de Khipu enviará al servidor del comercio cuando notifique que un pago está conciliado.",
            "example": "9dec8aa176c5223026919b3b5579a4776923e646ff3be686b9e6b62ec042e91f"
          },
          "receiver_id": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador único de una cuenta de cobro.",
            "example": 985101
          },
          "conciliation_date": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de conciliación del pago. Formato ISO-8601.",
            "example": "2017-03-01T13:00:00.000Z"
          },
          "subject": {
            "type": "string",
            "description": "Motivo del pago.",
            "example": "Test"
          },
          "amount": {
            "type": "number",
            "format": "double",
            "description": "El monto del cobro.",
            "example": 1000
          },
          "currency": {
            "type": "string",
            "description": "El código de moneda en formato ISO-4217.",
            "example": "CLP"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "verifying",
              "done"
            ],
            "description": "Estado del pago, puede ser `pending` (el pagador aún no comienza a pagar), `verifying` (se está verificando el pago) o `done`, cuando el pago ya está confirmado.",
            "example": "done"
          },
          "status_detail": {
            "type": "string",
            "enum": [
              "pending",
              "normal",
              "marked-paid-by-receiver",
              "rejected-by-payer",
              "marked-as-abuse",
              "reversed"
            ],
            "description": "Detalle del estado del pago:\n`pending` (el pagador aún no comienza a pagar),\n`normal` (el pago fue verificado y fue cancelado por algún medio de pago estándar),\n`marked-paid-by-receiver` (el cobrador marcó el cobro como pagado por otro medio),\n`rejected-by-payer` (el pagador declaró que no pagará),\n`marked-as-abuse` (el pagador declaró que no pagará y que el cobro fue no solicitado), y\n`reversed` (el pago fue anulado por el comercio, el dinero fue devuelto al pagador).",
            "example": "normal"
          },
          "body": {
            "type": "string",
            "description": "Detalle del cobro.",
            "example": "Test"
          },
          "picture_url": {
            "type": "string",
            "description": "URL con imagen del cobro.",
            "example": "https://micomercio.com/picture_url"
          },
          "receipt_url": {
            "type": "string",
            "description": "URL del comprobante de pago.",
            "example": "https://micomercio.com/order/receipt_url"
          },
          "return_url": {
            "type": "string",
            "description": "URL donde se redirige al pagador luego que termina el pago.",
            "example": "https://micomercio.com/order/return_url"
          },
          "cancel_url": {
            "type": "string",
            "description": "URL donde se redirige al pagador luego de que desiste hacer el pago.",
            "example": "https://micomercio.com/order/cancel_url"
          },
          "notify_url": {
            "type": "string",
            "description": "URL del webservice donde se notificará el pago.",
            "example": "https://micomercio.com/webhook/notify_url"
          },
          "notify_api_version": {
            "type": "string",
            "description": "Versión de la API de notificación.",
            "example": "3.0"
          },
          "expires_date": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha máxima para ejecutar el pago (en formato ISO-8601). El cliente podrá realizar varios intentos de pago hasta dicha fecha. Cada intento tiene un plazo individual de 3 horas para su ejecución.",
            "example": "2023-12-31T15:45:00-04:00"
          },
          "attachment_urls": {
            "type": "array",
            "minItems": 0,
            "description": "Arreglo de URLs de archivos adjuntos al pago.",
            "items": {
              "type": "string",
              "description": "URL de archivo adjunto al pago.",
              "example": "https://micomercio.com/attachment1.pdf"
            }
          },
          "bank": {
            "type": "string",
            "description": "Nombre del banco seleccionado por el pagador.",
            "example": "Banco de Chile (Edwards Citi)"
          },
          "bank_id": {
            "type": "string",
            "description": "Identificador del banco seleccionado por el pagador.",
            "example": "dfFbF"
          },
          "payer_name": {
            "type": "string",
            "description": "Nombre del pagador.",
            "example": "Nombre Pagador"
          },
          "payer_email": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico del pagador.",
            "example": "pagador@email.com"
          },
          "personal_identifier": {
            "type": "string",
            "description": "Identificador personal del pagador.",
            "example": "11.000.111-9"
          },
          "bank_account_number": {
            "type": "string",
            "description": "Número de cuenta bancaria del pagador.",
            "example": "001120490689"
          },
          "out_of_date_conciliation": {
            "type": "boolean",
            "description": "Es `true` si la conciliación del pago fue hecha luego de la fecha de expiración.",
            "example": true
          },
          "transaction_id": {
            "type": "string",
            "description": "Identificador del pago asignado por el cobrador.",
            "example": "zwo3wqz6uulcvajt"
          },
          "custom": {
            "type": "string",
            "description": "Campo genérico que asigna el cobrador al momento de hacer el pago.",
            "example": "<xml>...</xml>"
          },
          "responsible_user_email": {
            "type": "string",
            "description": "Correo electrónico de la persona responsable del pago.",
            "example": "responsible@email.com"
          },
          "send_reminders": {
            "type": "boolean",
            "description": "Es `true` cuando este es un cobro por correo electrónico y Khipu enviará recordatorios.",
            "example": true
          },
          "send_email": {
            "type": "boolean",
            "description": "Es `true` cuando Khipu enviará el cobro por correo electrónico.",
            "example": true
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "regular_transfer",
              "simplified_transfer",
              "not_available"
            ],
            "description": "Método de pago usado por el pagador, puede ser `regular_transfer` (transferencia normal) o `simplified_transfer` (transferencia simplificada).",
            "example": "simplified_transfer"
          },
          "funds_source": {
            "type": "string",
            "enum": [
              "debit",
              "prepaid",
              "credit",
              "not-available",
              ""
            ],
            "description": "Origen de fondos usado por el pagador, puede ser `debit` para pago con débito, `prepaid` para pago con prepago, `credit` para pago con crédito, o vacío en el caso de que se haya pagado mediante transferencia bancaria.",
            "example": "debit"
          },
          "discount": {
            "type": "number",
            "format": "double",
            "description": "Monto a descontar del valor pagado.",
            "example": 0
          },
          "third_party_authorization_details": {
            "type": "string",
            "description": "Ignorar este campo."
          }
        }
      },
...
    }
  }
}